Refactor to common storefinder
